pyproject.toml
[tool.poetry]
name = "quiz-bot"
version = "0.1.0"
description = ""
authors = ["ilyashirko <ilyashirko@gmail.com>"]
readme = "README.md"
packages = [{include = "quiz_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.14"
tqdm = "^4.64.1"
environs = "^9.5.0"
redis = "^4.3.4"
vk-api = "^11.9.9"
flake8 = "^5.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
